+++ irodsadmin/tests/test_commands.py
@@ -1,9 +1,10 @@
 """Management command tests for the irodsadmin app"""
 
-import io
 import os
 import sys
 import uuid
+
+from io import StringIO
 
 from irods.access import iRODSAccess
 from irods.path import iRODSPath
@@ -11,19 +12,14 @@ from irods.path import iRODSPath
 from django.conf import settings
 from django.core.management import call_command
 
-from test_plus.test import TestCase
-
 # Projectroles dependency
 from projectroles.constants import SODAR_CONSTANTS
 from projectroles.plugins import PluginAPI
-from projectroles.tests.test_models import (
-    ProjectMixin,
-    RoleMixin,
-    RoleAssignmentMixin,
-)
 
 # Landingzones dependency
+from landingzones.models import LandingZone
 from landingzones.tests.test_models import LandingZoneMixin
+from landingzones.tests.test_views_taskflow import LandingZoneTaskflowMixin
 
 # Samplesheets dependency
 from samplesheets.tests.test_io import SampleSheetIOMixin, SHEET_DIR
@@ -60,6 +56,7 @@ PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 
 # Local constants
 SHEET_PATH = SHEET_DIR + 'i_small.zip'
+SHEET_PATH_GERMLINE = os.path.join(SHEET_DIR, 'bih_germline.zip')
 ZONE_TITLE = '20180503_172456_test_zone'
 ZONE_DESC = 'description'
 DUMMY_UUID = '11111111-1111-1111-1111-111111111111'
@@ -67,6 +64,7 @@ USER_ADMIN = settings.IRODS_USER
 USER_NEW = 'user_new'
 LOGGER_PREFIX = 'irodsadmin.management.commands.'
 TEST_OBJ = 'test1.txt'
+OUT_SUFFIX = '0;0 bytes\n'
 
 
 class TestCheckSampleAccess(
@@ -212,56 +210,29 @@ class TestCheckSampleAccess(
         )
 
 
-# TODO: Modify this to use taskflow test base
 class TestIrodsOrphans(
     SampleSheetIOMixin,
+    SampleSheetTaskflowMixin,
     LandingZoneMixin,
-    RoleAssignmentMixin,
-    ProjectMixin,
-    RoleMixin,
-    TestCase,
+    LandingZoneTaskflowMixin,
+    TaskflowViewTestBase,
 ):
     """Tests for the irodsorphans management command"""
 
-    @staticmethod
-    def _get_stdout() -> str:
-        """Call irodsorphans management command and return output"""
-        out = io.StringIO()
-        sys.stdout = out
-        call_command('irodsorphans', stdout=out)
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
-        return output
+    def _setup_investigation(self, path: str = SHEET_PATH_GERMLINE):
+        """
+        Set up investigation with taskflow.
 
-    def setUp(self):
-        # Init roles
-        self.init_roles()
-        # Init super user
-        self.user = self.make_user('user')
-        self.user.is_superuser = True
-        self.user.is_staff = True
-        self.user.save()
-        # Init project with owner
-        self.project = self.make_project(
-            'TestProject',
-            PROJECT_TYPE_PROJECT,
-            None,
-        )
-        self.owner_as = self.make_assignment(
-            self.project, self.user, self.role_owner
-        )
-
-        self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
-        self.investigation.irods_status = True
-        self.investigation.save()
+        NOTE: This assumes a single study and a single assay
+        """
+        self.investigation = self.import_isa_from_file(path, self.project)
         self.study = self.investigation.studies.first()
         self.assay = self.study.assays.first()
-        self.assay.measurement_type = 'genome sequencing'
-        self.assay.technology_type = 'nucleotide sequencing'
-        self.assay.save()
+        self.make_irods_colls(self.investigation)
 
-        # Create LandingZone
-        self.landing_zone = self.make_landing_zone(
+    def _get_zone(self) -> LandingZone:
+        """Create and return landing zone for project with taskflow"""
+        zone = self.make_landing_zone(
             title=ZONE_TITLE,
             project=self.project,
             user=self.user,
@@ -270,428 +241,372 @@ class TestIrodsOrphans(
             configuration=None,
             config_data={},
         )
-        self.irods_backend = plugin_api.get_backend_api('omics_irods')
-        self.irods = self.irods_backend.get_session_obj()
+        self.make_zone_taskflow(zone)
+        return zone
 
-        # Create the actual assay, study and landing zone in the irods session
-        self.irods.collections.create(self.irods_backend.get_path(self.assay))
-        self.irods.collections.create(
-            self.irods_backend.get_path(self.landing_zone)
+    def _get_expected_colls(self) -> list:
+        """Return expected collections"""
+        return [
+            *self.command._get_assay_colls([self.assay]),
+            *self.command._get_study_colls([self.study]),
+            *self.command._get_zone_colls(),
+            *self.command._get_project_colls(),
+            *self.command._get_assay_subcolls([self.study]),
+        ]
+
+    @staticmethod
+    def _call_output() -> str:
+        """Call irodsorphans management command and return output"""
+        sys.stdout = StringIO()
+        call_command('irodsorphans', stdout=sys.stdout)
+        return sys.stdout.getvalue()
+
+    def setUp(self):
+        super().setUp()
+        self.project, self.owner_as = self.make_project_taskflow(
+            title='TestProject',
+            type=PROJECT_TYPE_PROJECT,
+            parent=self.category,
+            owner=self.user,
         )
+        self.command = Command()
 
-        # Set up the command
-        self.irodsorphans = Command()
-        self.expected_collections = (
-            *self.irodsorphans._get_assay_collections([self.assay]),
-            *self.irodsorphans._get_study_collections([self.study]),
-            *self.irodsorphans._get_zone_collections(),
-            *self.irodsorphans._get_project_collections(),
-            *self.irodsorphans._get_assay_subcollections([self.study]),
-        )
-
-    def tearDown(self):
-        self.irods.collections.get(
-            self.irods_backend.get_projects_path()
-        ).remove(force=True)
-        self.irods.cleanup()
-        super().tearDown()
-
-    def test_get_assay_collections(self):
-        """Test get_assay_collections()"""
+    def test_get_assay_colls(self):
+        """Test _get_assay_colls()"""
+        self._setup_investigation()
         self.assertListEqual(
-            self.irodsorphans._get_assay_collections([self.assay]),
+            self.command._get_assay_colls([self.assay]),
             [self.irods_backend.get_path(self.assay)],
         )
 
-    def test_get_study_collections(self):
-        """Test get_study_collections()"""
+    def test_get_study_colls(self):
+        """Test _get_study_colls()"""
+        self._setup_investigation()
         self.assertListEqual(
-            self.irodsorphans._get_study_collections([self.study]),
+            self.command._get_study_colls([self.study]),
             [self.irods_backend.get_path(self.study)],
         )
 
-    def test_get_zone_collections(self):
-        """Test get_zone_collections()"""
+    def test_get_zone_colls(self):
+        """Test _get_zone_colls()"""
+        self._setup_investigation()
+        zone = self._get_zone()
         self.assertListEqual(
-            self.irodsorphans._get_zone_collections(),
-            [self.irods_backend.get_path(self.landing_zone)],
+            self.command._get_zone_colls(),
+            [self.irods_backend.get_path(zone)],
         )
 
-    def test_get_project_collections(self):
-        """Test get_project_collections()"""
+    def test_get_project_colls(self):
+        """Test _get_project_colls()"""
+        self._setup_investigation()
         self.assertListEqual(
-            self.irodsorphans._get_project_collections(),
+            self.command._get_project_colls(),
             [self.irods_backend.get_path(self.project)],
         )
 
-    def test_get_assay_subcollections(self):
-        """Test get_assay_subcollections()"""
+    def test_get_assay_subcolls(self):
+        """Test _get_assay_subcolls()"""
+        self._setup_investigation()
         assay_path = self.irods_backend.get_path(self.assay)
+        expected = []
+        for prefix in ['p1', 'p2']:
+            for suffix in ['', '_mother', '_father']:
+                expected.append(
+                    iRODSPath(assay_path, f'{prefix}{suffix}-N1-DNA1-WES1')
+                )
+        expected.append(iRODSPath(assay_path, 'TrackHubs'))
+        expected.append(iRODSPath(assay_path, 'ResultsReports'))
+        expected.append(iRODSPath(assay_path, 'MiscFiles'))
         self.assertListEqual(
-            self.irodsorphans._get_assay_subcollections([self.study]),
-            [
-                iRODSPath(assay_path, '0815-N1-DNA1'),
-                iRODSPath(assay_path, '0815-T1-DNA1'),
-                iRODSPath(assay_path, 'TrackHubs'),
-                iRODSPath(assay_path, 'ResultsReports'),
-                iRODSPath(assay_path, 'MiscFiles'),
-            ],
+            self.command._get_assay_subcolls([self.study]),
+            expected,
         )
 
     def test_is_zone(self):
-        """Test is_zone()"""
-        collection = self.irods.collections.get(
-            self.irods_backend.get_path(self.landing_zone)
-        )
-        self.assertTrue(self.irodsorphans._is_zone(collection))
+        """Test _is_zone()"""
+        self._setup_investigation()
+        zone = self._get_zone()
+        coll = self.irods.collections.get(self.irods_backend.get_path(zone))
+        self.assertTrue(self.command._is_zone(coll))
 
     def test_is_assay_or_study_with_assay(self):
-        """Test is_assay_or_study() with assay"""
-        collection = self.irods.collections.get(
+        """Test _is_assay_or_study() with assay"""
+        self._setup_investigation()
+        coll = self.irods.collections.get(
             self.irods_backend.get_path(self.assay)
         )
-        self.assertTrue(self.irodsorphans._is_assay_or_study(collection))
+        self.assertTrue(self.command._is_assay_or_study(coll))
 
     def test_is_assay_or_study_with_study(self):
-        """Test is_assay_or_study() with study"""
-        collection = self.irods.collections.get(
+        """Test _is_assay_or_study() with study"""
+        self._setup_investigation()
+        coll = self.irods.collections.get(
             self.irods_backend.get_path(self.study)
         )
-        self.assertTrue(self.irodsorphans._is_assay_or_study(collection))
+        self.assertTrue(self.command._is_assay_or_study(coll))
 
     def test_is_project(self):
-        """Test is_project()"""
-        collection = self.irods.collections.get(
+        """Test _is_project()"""
+        coll = self.irods.collections.get(
             self.irods_backend.get_path(self.project)
         )
         projects_path = self.irods_backend.get_projects_path()
-        self.assertTrue(
-            self.irodsorphans._is_project(projects_path, collection)
-        )
+        self.assertTrue(self.command._is_project(projects_path, coll))
 
     def test_is_zone_invalid(self):
-        """Test is_zone() with a non-landingzone collection"""
-        collection = self.irods.collections.get(
+        """Test _is_zone() with a non-landingzone collection"""
+        coll = self.irods.collections.get(
             self.irods_backend.get_path(self.project)
         )
-        self.assertFalse(self.irodsorphans._is_zone(collection))
+        self.assertFalse(self.command._is_zone(coll))
 
     def test_is_assay_or_study_invalid(self):
-        """Test is_assay_or_study() with non-assay/study collection"""
-        collection = self.irods.collections.get(
+        """Test _is_assay_or_study() with non-assay/study collection"""
+        coll = self.irods.collections.get(
             self.irods_backend.get_path(self.project)
         )
-        self.assertFalse(self.irodsorphans._is_assay_or_study(collection))
+        self.assertFalse(self.command._is_assay_or_study(coll))
 
     def test_get_orphans_none(self):
-        """Test get_orphans() with no orphans available"""
+        """Test _get_orphans() with no orphans available"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
         # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
-        self.assertEqual(output, '')
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
+        self.assertEqual(sys.stdout.getvalue(), '')
 
     def test_get_orphans_assay(self):
-        """Test get_orphans() with orphan assay"""
-        orphan_path = '{}/assay_{}'.format(
-            self.irods_backend.get_path(self.study), uuid.uuid4()
+        """Test _get_orphans() with orphan assay"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.study), f'assay_{uuid.uuid4()}'
         )
         self.irods.collections.create(orphan_path)
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
         expected = (
-            str(self.project.sodar_uuid)
-            + ';'
-            + self.project.title
-            + ';'
-            + orphan_path
-            + ';0;0 bytes\n'
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
-        self.assertEqual(output, expected)
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
     def test_get_orphans_study(self):
-        """Test get_orphans() with orphan study"""
-        orphan_path = '{}/sample_data/study_{}'.format(
-            self.irods_backend.get_path(self.project), uuid.uuid4()
+        """Test _get_orphans() with orphan study"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.project),
+            settings.IRODS_SAMPLE_COLL,
+            f'study_{uuid.uuid4()}',
         )
         self.irods.collections.create(orphan_path)
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
         expected = (
-            str(self.project.sodar_uuid)
-            + ';'
-            + self.project.title
-            + ';'
-            + orphan_path
-            + ';0;0 bytes\n'
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
-        self.assertEqual(output, expected)
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
-    def test_get_output_zone(self):
-        """Test get_orphans() with orphan landing zone"""
-        collection = '20201031_123456'
-        orphan_path = '{}/landing_zones/{}/{}/{}'.format(
+    def test_get_orphans_zone(self):
+        """Test _get_orphans() with orphan landing zone collection"""
+        self._setup_investigation()
+        # Set up real zone to ensure only orphan is returned
+        self._get_zone()
+        expected_colls = self._get_expected_colls()
+        orphan_path = iRODSPath(
             self.irods_backend.get_path(self.project),
+            settings.IRODS_LANDING_ZONE_COLL,
             self.user.username,
             self.study.get_display_name().replace(' ', '_').lower(),
-            collection,
+            '20201031_123456',
         )
         self.irods.collections.create(orphan_path)
-
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
         expected = (
-            str(self.project.sodar_uuid)
-            + ';'
-            + self.project.title
-            + ';'
-            + orphan_path
-            + ';0;0 bytes\n'
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
-        self.assertEqual(output, expected)
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
     def test_get_output_project(self):
-        """Test get_orphans() with orphan project"""
-        collection = 'aa/aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa'
+        """Test _get_orphans() with orphan project"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
         orphan_path = iRODSPath(
-            self.irods_backend.get_projects_path(), collection
+            self.irods_backend.get_projects_path(), '11', DUMMY_UUID
         )
         self.irods.collections.create(orphan_path)
-
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
-        expected = (
-            collection[3:] + ';' + DELETED + ';' + orphan_path + ';0;0 bytes\n'
-        )
-        self.assertEqual(output, expected)
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
+        expected = f'{DUMMY_UUID};{DELETED};{orphan_path};{OUT_SUFFIX}'
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
     def test_get_output_assay_subs(self):
-        """Test get_orphans() with orphan assay subcollections"""
-        collection = 'UnexpectedCollection'
+        """Test _get_orphans() with orphan assay subcollections"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
         orphan_path = iRODSPath(
-            self.irods_backend.get_path(self.assay), collection
+            self.irods_backend.get_path(self.assay), 'UnexpectedCollection'
         )
         self.irods.collections.create(orphan_path)
-
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
         expected = (
-            str(self.project.sodar_uuid)
-            + ';'
-            + self.project.title
-            + ';'
-            + orphan_path
-            + ';0;0 bytes\n'
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};0;0 bytes\n'
         )
-        self.assertEqual(output, expected)
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
-    def test_get_output_deleted_project(self):
-        """Test get_output() with a deleted project"""
-        project_uuid = 'aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa'
-        collection = 'aa/' + project_uuid
-        orphan_path = '{}/{}'.format(
-            os.path.dirname(
-                os.path.dirname(self.irods_backend.get_path(self.project))
-            ),
-            collection,
+    def test_get_orphans_deleted_project(self):
+        """Test _get_orphans() with a deleted project"""
+        self._setup_investigation()
+        expected_colls = self._get_expected_colls()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_projects_path(), '11', DUMMY_UUID
         )
         self.irods.collections.create(orphan_path)
-
-        # Capture stdout
-        out = io.StringIO()
-        sys.stdout = out
-        self.irodsorphans._get_orphans(
-            self.irods,
-            self.expected_collections,
-            [self.assay],
-        )
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
-        expected = (
-            project_uuid + ';' + DELETED + ';' + orphan_path + ';0;0 bytes\n'
-        )
-        self.assertEqual(output, expected)
+        sys.stdout = StringIO()
+        self.command._get_orphans(self.irods, expected_colls, [self.assay])
+        expected = f'{DUMMY_UUID};{DELETED};{orphan_path};{OUT_SUFFIX}'
+        self.assertEqual(sys.stdout.getvalue(), expected)
 
     def test_command_no_orphans(self):
         """Test command with no orphans"""
-        out = io.StringIO()
-        sys.stdout = out
-        call_command('irodsorphans', stdout=out)
-        output = out.getvalue()
-        sys.stdout = sys.__stdout__
-        self.assertEqual(output, '')
+        self.assertEqual(self._call_output(), '')
 
     def test_command_orphan_assay(self):
         """Test command with orphan assay"""
-        orphan_path = '{}/assay_{}'.format(
-            self.irods_backend.get_path(self.study), uuid.uuid4()
+        self._setup_investigation()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.study), f'assay_{uuid.uuid4()}'
         )
         self.irods.collections.create(orphan_path)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)
 
     def test_command_orphan_study(self):
         """Test command with orphan study"""
-        orphan_path = '{}/sample_data/study_{}'.format(
-            self.irods_backend.get_path(self.project), uuid.uuid4()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.project),
+            settings.IRODS_SAMPLE_COLL,
+            f'study_{uuid.uuid4()}',
         )
         self.irods.collections.create(orphan_path)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)
 
     def test_command_orphan_zone(self):
-        """Test command with orphan landing zone"""
-        collection = '20201031_123456'
-        orphan_path = '{}/landing_zones/{}/{}/{}'.format(
+        """Test command with orphan landing zone collection"""
+        self._setup_investigation()
+        self._get_zone()
+        orphan_path = iRODSPath(
             self.irods_backend.get_path(self.project),
+            settings.IRODS_LANDING_ZONE_COLL,
             self.user.username,
             self.study.get_display_name().replace(' ', '_').lower(),
-            collection,
+            '20201031_123456',
         )
         self.irods.collections.create(orphan_path)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)
 
     def test_command_orphan_project(self):
         """Test command with orphan project"""
-        project_uuid = 'aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa'
-        collection = 'aa/' + project_uuid
         orphan_path = iRODSPath(
-            self.irods_backend.get_projects_path(), collection
+            self.irods_backend.get_projects_path(), '11', DUMMY_UUID
         )
         self.irods.collections.create(orphan_path)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            project_uuid, DELETED, orphan_path
-        )
+        output = self._call_output()
+        expected = f'{DUMMY_UUID};{DELETED};{orphan_path};{OUT_SUFFIX}'
         self.assertEqual(output, expected)
 
     def test_command_orphan_assay_sub(self):
         """Test command with orphan assay subcollection"""
-        collection = 'UnexpectedCollection'
-        orphan_path = '{}/{}'.format(
-            self.irods_backend.get_path(self.assay), collection
+        self._setup_investigation()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.assay), 'UnexpectedCollection'
         )
         self.irods.collections.create(orphan_path)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)
 
     def test_command_multiple(self):
         """Test command with multiple orphans"""
-        orphan_path = '{}/sample_data/study_{}'.format(
-            self.irods_backend.get_path(self.project), uuid.uuid4()
+        self._setup_investigation()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.project),
+            settings.IRODS_SAMPLE_COLL,
+            f'study_{uuid.uuid4()}',
         )
         self.irods.collections.create(orphan_path)
-        collection = '20201031_123456'
-        orphan_path2 = '{}/landing_zones/{}/{}/{}'.format(
+        orphan_path2 = iRODSPath(
             self.irods_backend.get_path(self.project),
+            settings.IRODS_LANDING_ZONE_COLL,
             self.user.username,
             self.study.get_display_name().replace(' ', '_').lower(),
-            collection,
+            '20201031_123456',
         )
         self.irods.collections.create(orphan_path2)
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path2,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path2};{OUT_SUFFIX}'
         )
-        expected += '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        expected += (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)
 
-    def test_command_ordering(self):
-        """Test ordering of orphans in command output"""
-        orphan_path = '{}/sample_data/study_{}'.format(
-            self.irods_backend.get_path(self.project), uuid.uuid4()
+    def test_command_output_order(self):
+        """Test command output order"""
+        self.maxDiff = None
+        self._setup_investigation()
+        orphan_path = iRODSPath(
+            self.irods_backend.get_path(self.project),
+            settings.IRODS_SAMPLE_COLL,
+            f'study_{uuid.uuid4()}',
         )
         self.irods.collections.create(orphan_path)
 
-        project2 = self.make_project('TestProject2', PROJECT_TYPE_PROJECT, None)
+        project2 = self.make_project(
+            'TestProject2', PROJECT_TYPE_PROJECT, self.category
+        )
         self.make_assignment(project2, self.user, self.role_owner)
-        orphan_path2 = '{}/sample_data/study_{}'.format(
-            self.irods_backend.get_path(project2), uuid.uuid4()
+        orphan_path2 = iRODSPath(
+            self.irods_backend.get_path(project2),
+            settings.IRODS_SAMPLE_COLL,
+            f'study_{uuid.uuid4()}',
         )
         self.irods.collections.create(orphan_path2)
 
-        output = self._get_stdout()
-        expected = '{};{};{};0;0 bytes\n'.format(
-            str(self.project.sodar_uuid),
-            self.project.full_title,
-            orphan_path,
+        output = self._call_output()
+        expected = (
+            f'{self.project.sodar_uuid};{self.project.full_title};'
+            f'{orphan_path};{OUT_SUFFIX}'
         )
-        expected += '{};{};{};0;0 bytes\n'.format(
-            str(project2.sodar_uuid),
-            project2.full_title,
-            orphan_path2,
+        expected += (
+            f'{project2.sodar_uuid};{project2.full_title};'
+            f'{orphan_path2};{OUT_SUFFIX}'
         )
         self.assertEqual(output, expected)